Separated the Options class from the Sustenet.cs file to comply with CodeFactor.
