+++ src/Sustenet.cs
@@ -18,81 +18,9 @@
 namespace Sustenet
 {
     using System;
-    using System.Collections.Generic;
-    using System.Timers;
-    using NDesk.Options;
     using Utils;
     using Transport;
     using System.Threading;
-
-    class Options
-    {
-        public class OptionsData
-        {
-            public bool client = false;
-            public int maxClients = 5000;
-            public bool cluster = false;
-            public bool master = false;
-        }
-
-        /// <summary>
-        /// Loads command-line arguments. Defaults --master if no other connection type is provided.
-        /// </summary>
-        /// <param name="args"></param>
-        /// <returns></returns>
-        public static OptionsData GetOptions(string[] args)
-        {
-            OptionsData data = new OptionsData();
-
-            // TODO: Add functionality for duplicates of each connection type.
-            OptionSet options = new OptionSet()
-            {
-                {
-                    "client:",
-                    "starts a client and waits for Connect() to be triggered.",
-                    v => {
-                        if(v != null)
-                        {
-                            int.TryParse(v, out data.maxClients);
-                        }
-
-                        data.client = true;
-                    }
-                },
-                {
-                    "cluster",
-                    "starts a cluster server and uses the config file to connect to a master server.",
-                    v => { data.cluster = true; }
-                },
-                {
-                    "master",
-                    "starts a master server, uses the config file to set it up, and waits for clusters and clients to connect.",
-                    v => { data.master = true; }
-                }
-            };
-
-            List<string> extra;
-            try
-            {
-                extra = options.Parse(args);
-
-                // If nothing is set, set master to true.
-                if(data.master == false && data.cluster == false && data.client == false)
-                {
-                    data.master = true;
-                }
-
-                return data;
-            }
-            catch(OptionException e)
-            {
-                Console.Write("Sustenet: ");
-                Console.WriteLine(e.Message);
-                Console.WriteLine("Try `sustenet --help' for more information.");
-                throw;
-            }
-        }
-    }
 
     class Sustenet
     {
